10NOV #1

+++ gongo/src/main/java/gongo/gongo/repository/ListProductRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import gongo.gongo.entity.ListProduct;
 
-public interface ListProductRepository extends JpaRepository<ListProduct, Integer>{
+public interface ListProductRepository extends JpaRepository<ListProduct, Long>{
     
 }
